backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
supabase==2.15.1
openai==1.75.0
requests==2.32.3
beautifulsoup4==4.13.3
python-dotenv==1.1.0
numpy==2.2.4
scipy==1.15.2
scikit-learn==1.6.1